refactor: portals are separate now
