+++ src/components/autocomplete/autocomplete.tsx
@@ -1,6 +1,8 @@
-import styled from 'styled-components'
 import React from 'react'
+import styled from 'styled-components'
+import { RawDropdown } from '../dropdown'
 import { AutocompleteState } from './use-autocomplete'
+import { useDisclosure } from '../../contexts/disclosure'
 
 const AutocompleteContainer = styled.div`
   position: relative;
@@ -69,17 +71,23 @@ const onChange = (state: AutocompleteState, value: string) => {
  */
 export const Autocomplete = (props: AutocompleteProps) => {
   console.info(`${JSON.stringify(props)}`)
+  const disclosure = useDisclosure(false)
+  const portalContent = (
+    <AutocompleteSuggestions>
+      <AutocompleteSuggestionItem>USA</AutocompleteSuggestionItem>
+      <AutocompleteSuggestionItem>India</AutocompleteSuggestionItem>
+      <AutocompleteSuggestionItem>Argentina</AutocompleteSuggestionItem>
+    </AutocompleteSuggestions>
+  )
   return (
-    <AutocompleteContainer>
-      <AutocompleteInput
-        placeholder={props.placeholder}
-        onChange={(e) => onChange(props.state, e.target.value)}
-      />
-      <AutocompleteSuggestions>
-        <AutocompleteSuggestionItem>USA</AutocompleteSuggestionItem>
-        <AutocompleteSuggestionItem>India</AutocompleteSuggestionItem>
-        <AutocompleteSuggestionItem>Argentina</AutocompleteSuggestionItem>
-      </AutocompleteSuggestions>
-    </AutocompleteContainer>
+    <RawDropdown disclosure={disclosure} portalContent={portalContent}>
+      <AutocompleteContainer>
+        <AutocompleteInput
+          onClick={() => disclosure.toggleOpen()}
+          placeholder={props.placeholder}
+          onChange={(e) => onChange(props.state, e.target.value)}
+        />
+      </AutocompleteContainer>
+    </RawDropdown>
   )
 }
