+++ src/components/dropdown/rawdropdown.test.tsx
@@ -1,0 +1,7 @@
+import { RawDropdown } from './rawdropdown'
+
+describe('RawDropdown', () => {
+  it('is truthy', () => {
+    expect(RawDropdown).toBeTruthy()
+  })
+})
